Updated the chunking logic to use Pyannote v3 for VAD to intelligently chunk the audio based on no speech detection windows to avoid hitting hard fallback cuts.
TODO: In 2 of 14 test cases, timestamp drift occurred when non-speech segments were ~30-60 seconds long. This made the subtities start far earlier than they should have which impacted the timing of the remainder of the subtitles for that chunk because they'd also be offset by the same incorrect timing. More investigation is needed to detect and prevent this from happening.